+++ app.py
@@ -1,9 +1,12 @@
-from flask import Flask
+from flask import Flask, url_for
 from flask import render_template
 from flask import request
+from flask import redirect
+from flask import url_for
 import mysql.connector
 import json
 import regex
+import database
 
 app = Flask(__name__)
 
@@ -13,63 +16,23 @@ def form():
 
 @app.route('/handle_data', methods=['POST'])
 def handle_data():
-    name = request.form["name"]
-    email = request.form["email"]
-    if(regex.checkEmail(email)):
-      return "success"
-    else:
-      return "fail"
+  name = request.form["name"]
+  email = request.form["email"]
+    
+  if(regex.checkEmail(email)):
+    database.db_init()
+    database.insertIntoDatabase(name, email)
+  else:
+    return "fail"
 
-@app.route('/widgets')
-def get_widgets():
-  mydb = mysql.connector.connect(
-    host="mysqldb",
-    user="root",
-    password="p@ssw0rd1",
-    database="inventory"
-  )
-  cursor = mydb.cursor()
+  return redirect('/')
+    
 
 
-  cursor.execute("SELECT * FROM widgets")
+@app.route('/view')
+def viewDatabase():
+  return database.view()
 
-  row_headers=[x[0] for x in cursor.description] #this will extract row headers
-
-  results = cursor.fetchall()
-  json_data=[]
-  for result in results:
-    json_data.append(dict(zip(row_headers,result)))
-
-  cursor.close()
-
-  return json.dumps(json_data)
-
-@app.route('/initdb')
-def db_init():
-  mydb = mysql.connector.connect(
-    host="mysqldb",
-    user="root",
-    password="p@ssw0rd1"
-  )
-  cursor = mydb.cursor()
-
-  cursor.execute("DROP DATABASE IF EXISTS inventory")
-  cursor.execute("CREATE DATABASE inventory")
-  cursor.close()
-
-  mydb = mysql.connector.connect(
-    host="mysqldb",
-    user="root",
-    password="p@ssw0rd1",
-    database="inventory"
-  )
-  cursor = mydb.cursor()
-
-  cursor.execute("DROP TABLE IF EXISTS widgets")
-  cursor.execute("CREATE TABLE widgets (name VARCHAR(255), description VARCHAR(255))")
-  cursor.close()
-
-  return 'init database'
 
 
 if __name__ == "__main__":
